Remove dependency on YourVote useEffect
This was causing an infinte loop. The proper solution would be to
refactor useFetch so that it's wrapping fetchMethod in a useCallback,
though I'm too tired to do that work now/it may effect other parts of
the project as useFetch is used everywhere.

Going to leave this up to whoever takes the ticket to fix all the warnings.